Update configure-edge-with-intune.md
The doc needs to be corrected.
In the heading and in the 1st line we are mentioning "Administrative Templates" however just below that the doc explains how to set the policy using Settings Catalog. The screenshots that follow are also from settings catalog.
Hence "Administrative template" needs to be corrected with Settings Catalog in this doc.